move codegen parts of type.cpp to codegen/
